Shellmet

Description

This module contains neat utilities to be able to work with shell commands in generic and simple way using just string literals.

>>> "echo" ["Hello", "World!"]
⚙  echo Hello 'World!'
Hello World!

Documentation

($|) :: FilePath -> [Text] -> IO Text infix 5

Run shell command with given options and return stripped stdout of the executed command.

>>> "echo" $| ["Foo", "Bar"]
"Foo Bar"

($^) :: FilePath -> [Text] -> IO () infix 5

This operator runs shell command with given options but doesn't print the command itself.

>>> "echo" $^ ["Foo", "Bar"]
Foo Bar

($?) :: IO a -> IO a -> IO a infixl 4

Do some IO actions when process failed with IOError.

>>> "echo" ["0"] $? putStrLn "Command failed"
⚙  echo 0
0

>>> "exit" ["1"] $? putStrLn "Command failed"
⚙  exit 1
Command failed

isSuccess :: IO a -> IO Bool

Returns the indicator of if the command succeded or not.

>>> isSuccess $ "echo" ["Hello world!"]
⚙  echo 'Hello world!'
Hello world!
True

Orphan instances

(a ~ [Text], b ~ IO ()) => IsString (a -> b)
This instance is needed to provide functionality to call commands by using simple string literals in IO monad.

>>> "ls" ["-1", "test"]
⚙  ls -1 test
Doctest.hs

Instance details

Methods

fromString :: String -> a -> b